refactor: implement clean architecture on app
