+++ .prettierrc
@@ -1,5 +1,6 @@
 {
   "tabWidth": 2,
   "useTabs": false,
-  "singleQuote": true
+  "singleQuote": true,
+  "bracketSameLine": true
 }
