+++ src/infrastructure/index.js
@@ -1,1 +1,2 @@
-export { default as api } from './api';
+export { default as api } from './services/api';
+export { default as isTextMatch } from './services/search';
