+++ src/infrastructure/services/api.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { FETCH_SHOWS_URL } from '../constants';
+import { FETCH_SHOWS_URL } from '../../constants';
 
 const api = {
   getShows: async () => {
