+++ src/infrastructure/services/search.js
@@ -1,0 +1,6 @@
+function isTextMatch(item, word) {
+  const regex = new RegExp(word, 'ig');
+  return item.name.match(regex);
+}
+
+export default isTextMatch;
